// packages/sd-orm/src/common/Enums.ts
export enum OrderByRule {
  ASC = "ASC",
  DESC = "DESC"
}

export enum IndexType {
  DEFAULT = "",
  UNIQUE = "UNIQUE",
  FULLTEXT = "FULLTEXT"
}

export enum DataType {
  BIT = "BIT",
  INT = "INT",
  NVARCHAR = "NVARCHAR",
  DATETIME = "DATETIME",
  DATE = "DATE",
  TIME = "TIME",
  UNIQUEIDENTIFIER = "UNIQUEIDENTIFIER",
  VARBINARY = "VARBINARY"
}